# Improve the repo metadata generation implementation by caching the hash