--- a/llvm/include/llvm/IR/RepoHashCalculator.h
+++ b/llvm/include/llvm/IR/RepoHashCalculator.h
@@ -178,9 +178,7 @@ public:
 
   repodefinition::GOVec &getDependencies() { return Dependencies; }
 
-  repodefinition::GOVec &getContributedToGVs() {
-    return ContributedToGVs;
-  }
+  repodefinition::GOVec &getContributedToGVs() { return ContributedToGVs; }
 
 private:
   // Accumulate the hash of basicblocks, instructions and variables etc in the