… llvm/test/Feature/Repo/repo_hash_self_loop.ll
… llvm/test/Feature/Repo/repo_hash_simple_case.ll
… llvm/unittests/IR/RepoDefinitionTest.cpp