Two ways to improve the RepoMetadataGenerationPass execution time:

1) Separating the ‘visited’ table for the two directions. In other words, the contributions get a visited map and the dependencies get a separate map. This will stop the algorithm from a) thinking that it’s detecting lots of loops and b) accumulating unnecessary successors.
2) Caching the digests of each possible GOs to avoid the recalculation. It will give the further performance benefits.

The upstream LLVM+Clang is used as an application to check the performance. The performance improvement is summarised:
1) After applying the patch 1), the LLVM backend time for the first time build changed from 10269.5s to 9395.5s (8.5%improvement). The backend time for the second time build changed from 1913.5s to 1013.0s (47.06% improvement).
2) After applying the patch 2), the LLVM backend time for the first time build changed from 9395.5s to 9072.5s (further 34.38 %improvement). The backend time for the second time build changed from 1013.0s to 791.0s (further 21.91% improvement).

The RepoMetadataGenerationPass execution time (user time) of the SemaDeclAttr.cpp file has been reduced from 52,384.825ms to 5,092.376ms (90.28% improvement).

Fix for <https://github.com/SNSystems/llvm-project-prepo/issues/34>