@MaggieYingYi 2020-04-23
> 
> 
> Is it possible to split this into the two patches you mention in the description? It would be helpful if we ever want to go back and recreate those figures. One pull request and two commits would be okay from my point of view.

Good point. I have separated the patch into the two commits.