@rgal 2020-04-28
The first part is “Separating the ‘visited’ table for the two directions”, yet the code change (https://github.com/SNSystems/llvm-project-prepo/pull/86/commits/c5d3181bbfe9f4cd198eb180c52a3b2d94f0d206) doesn’t touch the “Visited” StateMap. Is this just because it  changes the order that the relationships are traversed. “the dependencies get a separate map”, I can’t see where this is implemented.