@MaggieYingYi 2020-04-30
As agreed by the reviewers, I have put patch one (c5d3181bb) to a separate [PR#89](https://github.com/SNSystems/llvm-project-prepo/pull/89) for code review.