@paulhuggett 2020-05-25
Commit paulhuggett/digraph-hash@97a44de9226acdc632416db40f2ebfbca14750da correct a bug in the algorithm that could result in fewer objects being cached. I’d recommend integrating that fix into this code.