@MaggieYingYi 2020-05-26
> 
> 
> Commit [paulhuggett/digraph-hash@97a44de](https://github.com/paulhuggett/digraph-hash/commit/97a44de9226acdc632416db40f2ebfbca14750da) correct a bug in the algorithm that could result in fewer objects being cached. I’d recommend integrating that fix into this code.

Yes, I have integrated in https://github.com/SNSystems/llvm-project-prepo/pull/86/commits/52aeb386.