@MaggieYingYi 2020-05-28
Thanks both. I will update the commit message and merge this patch back to master branch once I get the performance result. 
Regards,